+++ csrc/kernels.cuh
@@ -105,7 +105,7 @@ template<typename T, int OPTIMIZER, int BLOCK_SIZE, int N_PER_TH> __global__ voi
 template<typename T, int BLOCK_SIZE, int NUM_VALS> __global__ void kPercentileClipping(T * __restrict__ g, float *gnorm_vec, int step, const int n);
 
 __global__ void kHistogramScatterAdd2D(float* histogram, int *index1, int *index2, float *src, const int maxidx1, const int n);
-template <typename T, int FUNC> __global__ void kfunc(T *A, T value, int n);
+template <typename T, int FUNC> __global__ void kfunc(T *A, T value, long n);
 
 #endif
